requirements.txt
Flask
python-magic
Pillow
